# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SecureAI - Biometric Security Demo API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List all payments
      responses:
        "200":
          description: Payments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"

    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a simulated payment (Stripe-style tokenisation)
      description: Simulates a real processor decision (lib/paymentSimulation.ts) rather than always succeeding — a known Stripe test-card last4 (via cardLast4 in the body) can simulate a realistic decline. Supports the Idempotency-Key header, standard practice for payment APIs — a retried request with the same key returns the original payment instead of creating a duplicate charge.
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
          description: Optional client-generated key (e.g. a UUID) scoping this request as a single logical attempt. Retrying with the same key returns the original payment unchanged.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment created (or the original payment, if this Idempotency-Key was already used)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Simulated processor decline — see the payment's own declineCode/declineMessage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payments/{id}/refund:
    post:
      operationId: refundPayment
      tags: [payments]
      summary: Refund a completed payment
      description: Self-service within a 14-day window for the payment's own owner; admins can refund any completed payment regardless of age. Mirrors the same status transition a real payment.refunded webhook event would apply (routes/webhooks.ts), but attributed to a direct user/admin action rather than the (simulated) processor.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 2147483647
      responses:
        "200":
          description: Payment refunded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "400":
          description: Not eligible for refund (wrong status, or outside the self-service window)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get a payment by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 2147483647
      responses:
        "200":
          description: Payment found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /payments/subscribe:
    post:
      operationId: subscribe
      tags: [payments]
      summary: Subscribe to a plan — the server looks up the canonical price by planId, the client never sends an amount
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
          description: Same semantics as POST /payments — retrying with the same key returns the original subscription payment.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscribeInput"
      responses:
        "201":
          description: Subscription payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscribeResult"
        "400":
          description: Unknown plan or already subscribed to it
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Payment:
      type: object
      required: [id, userId, amount, currency, status, description, providerToken, createdAt]
      properties:
        id:
          type: number
        userId:
          description: Null if the owning account has since been deleted — the payment record is retained (userEmail preserves who it was), the account is not.
          type: ["number", "null"]
        userEmail:
          type: ["string", "null"]
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [pending, completed, failed, refunded]
        description:
          type: string
        declineCode:
          description: Set only when status is "failed" — see lib/paymentSimulation.ts. Null otherwise.
          type: ["string", "null"]
        declineMessage:
          description: A user-facing explanation of the decline. Null unless status is "failed".
          type: ["string", "null"]
        providerToken:
          type: string
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [amount, currency, description]
      properties:
        amount:
          description: No real payment processor is behind this demo endpoint, so there's no upstream cap enforcing a sane ceiling the way a real processor would — this one is Team 1's own, not a business decision to defer.
          type: number
          minimum: 0.01
          maximum: 999999.99
        currency:
          description: Restricted to the currencies this app actually prices plans in, not just "any 3 letters" — a real ISO-4217 registry check is out of scope for a demo, but "looks like a real currency" is not.
          type: string
          enum: [USD, EUR, GBP, AUD, CAD]
        description:
          type: string
          minLength: 1
        cardLast4:
          description: Optional. The last 4 digits only — never the full card number, expiry, or CVV, which never leave the browser (see cardValidation.ts). Used purely to drive the simulated processor's decline logic (lib/paymentSimulation.ts) against Stripe's own published test-card numbers; omitting it always simulates success.
          type: ["string", "null"]
          pattern: "^\\d{4}$"
        cardBrand:
          description: Optional, display-only — e.g. "Visa", "Mastercard". Not itself sensitive (a standard, publicly documented numbering scheme).
          type: ["string", "null"]
...
    SubscribeInput:
      type: object
      required: [planId]
      properties:
        planId:
          description: One of the Plan.id values from GET /payments/plans — never a client-supplied amount
          type: string
          minLength: 1
        cardLast4:
          description: Optional, same simulated-decline purpose as PaymentInput.cardLast4 — omit to always simulate success.
          type: ["string", "null"]
          pattern: "^\\d{4}$"
        cardBrand:
          type: ["string", "null"]
    SubscribeResult:
      type: object
      required: [payment, subscriptionPlan]
      properties:
        payment:
          $ref: "#/components/schemas/Payment"
        subscriptionPlan:
          type: string
          enum: [free, plus, pro, team]
...